change made

+++ src/main/java/np/edu/nast/payroll/Payroll/service/LeaveBalanceService.java
@@ -4,11 +4,11 @@ import np.edu.nast.payroll.Payroll.entity.LeaveBalance;
 import java.util.List;
 
 public interface LeaveBalanceService {
-
     LeaveBalance createLeaveBalance(LeaveBalance balance);
     LeaveBalance updateLeaveBalance(Long id, LeaveBalance balance);
     void deleteLeaveBalance(Long id);
     LeaveBalance getLeaveBalanceById(Long id);
     List<LeaveBalance> getAllLeaveBalances();
-    List<LeaveBalance> getLeaveBalanceByEmployee(Long empId);
+    // FIX: Change Long to Integer to align with Repository
+    List<LeaveBalance> getLeaveBalanceByEmployee(Integer empId);
 }
